pyproject.toml
[project]
name = "etlpipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.0.4",
    "apache-airflow-providers-http>=5.3.3",
    "apache-airflow-providers-postgres>=6.2.3",
    "mlflow<3.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "scikit-learn>=1.7.1",
]
